Further CSS added
Added another css file, this one is for the body of the page. Also updated the header css file to correctly apply to the current header format.

These CSS files will need to be changed in the future the colour scheme and layout is very ugly, the current CSS serves as a placeholder